requirements.txt
matplotlib==3.5.3
mmcv==0.2.14
numpy==1.21.6
opencv-python==4.6.0.66
pandas==1.3.5
pathlib==1.0.1
pathtools==0.1.2
Pillow==9.2.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
timm==0.6.11
tokenizers==0.12.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==0.11.4
torchtext==0.12.0
torchvision==0.12.0+cu113
tqdm==4.65.0
vidaug @ git+https://github.com/okankop/vidaug@1c1ddf2640fe4a9171267d64ae5e3bd70c24d54a
wandb==0.12.21
